Retire F2 and F3 Ranking retirement
If we want to put a F2 or a F3 driver in a F1 car, he has to keep hes ranking in his origin category. 
If you transfer a F2 or a F3 driver in seanson, thoses lines will deletes the ranking of the transfered driver, and will create a crash.

Tested out a game without thoses lines and it works. So. Maybe it'll be good if we delete it ?